Refactor name lookup logic out of Context
