+++ toolchain/check/BUILD
@@ -36,6 +36,7 @@ cc_library(
         "merge.cpp",
         "modifiers.cpp",
         "name_component.cpp",
+        "name_lookup.cpp",
         "operator.cpp",
         "pattern_match.cpp",
         "return.cpp",
@@ -67,6 +68,7 @@ cc_library(
         "merge.h",
         "modifiers.h",
         "name_component.h",
+        "name_lookup.h",
         "operator.h",
         "param_and_arg_refs_stack.h",
         "pattern_match.h",
