+++ toolchain/check/literal.cpp
@@ -7,6 +7,7 @@
 #include "toolchain/check/call.h"
 #include "toolchain/check/context.h"
 #include "toolchain/check/convert.h"
+#include "toolchain/check/name_lookup.h"
 
 namespace Carbon::Check {
 
@@ -22,8 +23,8 @@ auto MakeIntTypeLiteral(Context& context, Parse::NodeId node_id,
                         SemIR::IntKind int_kind, IntId size_id)
     -> SemIR::InstId {
   auto width_id = MakeIntLiteral(context, node_id, size_id);
-  auto fn_inst_id = context.LookupNameInCore(
-      node_id, int_kind == SemIR::IntKind::Signed ? "Int" : "UInt");
+  auto fn_inst_id = LookupNameInCore(
+      context, node_id, int_kind == SemIR::IntKind::Signed ? "Int" : "UInt");
   return PerformCall(context, node_id, fn_inst_id, {width_id});
 }
